A span is two u32, and the AST is eight bytes narrower for it (#1135)
* a span is two u32, and the AST is eight bytes narrower for it

`Span` held two `usize`. Every `Expr`, every `Stmt` and every diagnostic
carries one, so the width of that struct sets the width of the AST:

    Span     16 -> 8
    Expr     64 -> 56
    Stmt    136 -> 120
    Pattern  64 -> 64   (its largest arm is not span-bound)

The lexer and the passes still count in `usize`, so the narrowing happens
at one place — `Span::at(line, col)` — and `render` widens back for the
source-line lookup.

Measured on `kanso check lib/json`:

    compile_peak_bytes    822,004 -> 763,868    -58,136   -7.1%
    compile_instructions  58,329,928 -> 58,379,986  +50,058  +0.086%
    compile_allocs         61,974 ->  61,974    unchanged

`compile_peak_bytes` reads the same here as in the runner's golden, so
that counter is host-invariant and the fall is real rather than this
box's layout. Allocation count holds because the same objects are
allocated; they are smaller.

The instruction rise is the cast at the boundary: about twenty side
tables key on `(String, usize, usize)`, so a narrowed span is widened to
build a key and narrowed again to store one. Carrying `u32` through those
tables would recover most of it; that is a larger change than this one.

Welfare: 84.10 -> 84.29, +0.19.


* the profile does not localise the 50,058, and the padding probe says why

The entry claimed the instruction rise was the cast at the side-table
boundary. That was inferred, not measured, and measuring it does not
support it.

Two probes. A build with `u32` fields and two words of explicit padding
— same casts, same tables, `Span` back at 16 and `Expr` back at 64 —
reads `compile_peak_bytes=822004`, the baseline to the byte. So the
memory fall is the struct size and nothing else, which the entry can now
say as a fact.

The instructions do not resolve. Callgrind puts the unpadded build's rise
almost entirely in `walk_children` (+343k gross against falls elsewhere),
which does no span arithmetic; the padded build rises by the same amount
overall, +47,430, and attributes it to `parse_atom` and `parse_stmt`
with `walk_children` unchanged. Same total, disjoint explanations — the
signature of code layout rather than work. The runner's number decides.


* regenerate both compile goldens from the runner, and bank the gain

The container cannot measure either row — both are pinned to
rustc=1.98.0 and this box is 1.94.1 — so CI measured them:

    compile_peak_bytes   822,004 -> 763,868       -58,136   -7.1%
    compile_instructions 57,822,766 -> 58,205,543  +382,777  +0.66%
    compile_allocs       61,974 -> 61,974          unchanged

The memory row reads 763,868 on both hosts, which is what made the
container's measurement trustworthy in the first place.

The instruction row does not agree across hosts about its size: +382,777
on the runner against +50,058 here, same sign and an eighth the
magnitude. Some of it is work and some is layout, and the entry says so
rather than picking one. The goldens carry the runner's rows, the tagged
figure on docs/compiler.html carries the same number, and the welfare
floor moves to 84.27 with the reason attached.

The entry also records the next eight bytes, measured and left
unshipped: `Guard.rest` as a `Box<[Stmt]>` takes `Expr` to 48 and peak to
721,652, another 5.5%, for +2.1% instructions that land on every walker
matching an `Expr`. Net +0.08 welfare, which does not buy the encoding
question.


---------

+++ src/beat.rs
@@ -382,7 +382,11 @@ fn is_chain(
             Expr::Ident(n, _)
                 if matches!(n.as_str(), "append" | "builtin_append")
                     && args.len() == 2
-                    && mut_sites.contains(&(decl.file.clone(), span.line, span.col)) =>
+                    && mut_sites.contains(&(
+                        decl.file.clone(),
+                        span.line as usize,
+                        span.col as usize,
+                    )) =>
             {
                 is_chain(&args[0], own, decl, locals, mut_sites, chains, folds)
             }
@@ -1303,7 +1307,11 @@ fn map_chain_rest(
             Expr::Ident(n, _)
                 if matches!(n.as_str(), "put" | "builtin_put")
                     && args.len() == 3
-                    && mut_sites.contains(&(decl.file.clone(), span.line, span.col)) =>
+                    && mut_sites.contains(&(
+                        decl.file.clone(),
+                        span.line as usize,
+                        span.col as usize,
+                    )) =>
             {
                 literal_key(&args[1])
                     && scalar_elem(&args[2], program, own, decl, inference, decl_index)
@@ -1356,7 +1364,11 @@ fn is_scalar_list_chain(
             Expr::Ident(n, _)
                 if matches!(n.as_str(), "push" | "builtin_push")
                     && args.len() == 2
-                    && mut_sites.contains(&(decl.file.clone(), span.line, span.col)) =>
+                    && mut_sites.contains(&(
+                        decl.file.clone(),
+                        span.line as usize,
+                        span.col as usize,
+                    )) =>
             {
                 scalar_elem(&args[1], program, own, decl, inference, decl_index)
                     && is_scalar_list_chain(
